+++ landingzones/views.py
@@ -11,7 +11,6 @@ from django.urls import reverse
 from django.views.generic import TemplateView, CreateView
 
 # Projectroles dependency
-from projectroles.models import Project
 from projectroles.plugins import get_backend_api
 from projectroles.views import (
     LoggedInPermissionMixin,
@@ -31,7 +30,7 @@ from landingzones.forms import LandingZoneForm
 from landingzones.models import (
     LandingZone,
     STATUS_ALLOW_UPDATE,
-    STATUS_ALLOW_CLEAR,
+    STATUS_FINISHED,
 )
 
 
@@ -391,10 +390,15 @@ class ProjectZoneView(
         context['irods_webdav_enabled'] = int(settings.IRODS_WEBDAV_ENABLED)
         if settings.IRODS_WEBDAV_ENABLED:
             context['irods_webdav_url'] = settings.IRODS_WEBDAV_URL.rstrip('/')
+
         # User zones
-        context['zones_own'] = LandingZone.objects.filter(
-            project=context['project'], user=self.request.user
-        ).order_by('title')
+        context['zones_own'] = (
+            LandingZone.objects.filter(
+                project=context['project'], user=self.request.user
+            )
+            .exclude(status__in=STATUS_FINISHED)
+            .order_by('title')
+        )
 
         # Other zones
         # TODO: Add individual zone perm check if/when we implement issue #57
@@ -404,12 +408,12 @@ class ProjectZoneView(
             context['zones_other'] = (
                 LandingZone.objects.filter(project=context['project'])
                 .exclude(user=self.request.user)
+                .exclude(status__in=STATUS_FINISHED)
                 .order_by('user__username', 'title')
             )
 
         # Status query interval
         context['zone_status_interval'] = settings.LANDINGZONES_STATUS_INTERVAL
-
         return context
 
 
@@ -652,67 +656,3 @@ class ZoneMoveView(
             messages.error(self.request, str(ex))
 
         return redirect(redirect_url)
-
-
-class ZoneClearView(
-    LoginRequiredMixin,
-    LoggedInPermissionMixin,
-    ProjectContextMixin,
-    ProjectPermissionMixin,
-    TemplateView,
-):
-    """LandingZone validation and moving triggering view"""
-
-    http_method_names = ['get', 'post']
-    template_name = 'landingzones/landingzone_confirm_clear.html'
-    permission_required = 'landingzones.update_zones_own'
-
-    def post(self, request, **kwargs):
-        timeline = get_backend_api('timeline_backend')
-        project = Project.objects.get(sodar_uuid=self.kwargs['project'])
-        tl_event = None
-
-        # Add event in Timeline
-        if timeline:
-            tl_event = timeline.add_event(
-                project=project,
-                app_name=APP_NAME,
-                user=self.request.user,
-                event_name='zones_clear',
-                description='clear inactive landing zones from {user}',
-            )
-            tl_event.add_object(
-                obj=self.request.user,
-                label='user',
-                name=self.request.user.username,
-            )
-
-        try:
-            inactive_zones = LandingZone.objects.filter(
-                project=project,
-                user=self.request.user,
-                status__in=STATUS_ALLOW_CLEAR,
-            )
-            zone_count = inactive_zones.count()
-            inactive_zones.delete()
-            messages.success(
-                self.request,
-                'Cleared {} inactive landing zone{} for user {}.'.format(
-                    zone_count,
-                    's' if zone_count != 1 else '',
-                    self.request.user.username,
-                ),
-            )
-            if tl_event:
-                tl_event.set_status('OK')
-        except Exception as ex:
-            messages.error(
-                self.request,
-                'Unable to clear inactive landing zones: {}'.format(ex),
-            )
-            if tl_event:
-                tl_event.set_status('FAILED', str(ex))
-
-        return redirect(
-            reverse('landingzones:list', kwargs={'project': project.sodar_uuid})
-        )